moved code from core/displayhook.py to terminal/prompts.py
